requirements.txt
streamlit>=1.20
gspread>=5.8.0
google-auth>=2.20.0
requests>=2.31.0
pandas>=2.0.0
python-dateutil>=2.8.2
Pillow>=10.0.0
crcmod
qrcode
